requirements.txt
# Django and DRF
Django==3.2.7
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.1.0
django-money==2.1
Pillow==8.3.2
